Rename LocationManager to LocationRequestViewModel
